edited the original files
Once rendered, all except the CoAbundantTaxa.rmd (experimental) should be linked to the website. CoAbundantTaxa.rmd can be added as a new topic within the Microbiome analysis section but I leave the final decision for Leo. Plus it is better to move CompositionAmplicondata.Rmd, CompositionAmplicondata2.Rmd, CoremicrobiotaAmplicon.Rmd and respective *.html files to some another folder as these are old and not useful now.